Revise test cases for random test
To support random test, the semantics of mce-inject is changed. The
default injecting context is RANDOM instead of PROCESS. The test cases
are revised to add as much randomization as possible. Such as, the main
MCE is injected in RANDOM context, while one of the broadcasted MCE is
injected in PROCESS context.