tinjpage: make file clean work for non-NFS fs
The MAP_POPULATE must be removed to make file_clean work properly
in local filesystems (mine is ext4). When MAP_POPULATE is present
together with PROT_WRITE, the kernel side will (mistakenly?) set
the dirty bit:
                if (page_mkclean(p))
                        SetPageDirty(p);
Because page_mkclean() returns true for (pte_dirty || pte_write).

This could be a kernel bug, but for now let's avoid it in user space.

Signed-off-by: Wu Fengguang <fengguang.wu@intel.com>
Signed-off-by: Andi Kleen <ak@linux.intel.com>