Fix panic type and exception message for the new cases of recoverable_ucr
The new cases are srao_ewb_noripv and srao_mem_scrub_noripv. After either
one of them injected, the result of the case seems right. But panic type
and exception message needed as standards for checking the result of the
cases does not make sense in fact. Because they are "NULL". So I modify them.

Signed-off-by: Jin Dongming <jin.dongming@np.css.fujitsu.com>