Clear existed ERST record before the test
If ERST table is full, the test can't begin. To avoid
this potential issue, if existing ERST record, erase
one record to relase the storage space and let the test
go on.

Because the ERST test maybe damges the data in the ERST
table, please restore the valid data in the ERST to the
other safe place.

Signed-off-by: Chen Gong <gong.chen@linux.intel.com>