Avoid optimizing out the access to the injected location by compiler
The "SRAR DCU" test case failed on a CLX-AP server. It's root caused
that the gcc v8.2.1 optimized out the access to the injected location.

Move the "total" from being a local variable to a global to avoid the
optimization by the compiler. Consequently, it can make sure the poisoned
data is consumed to trigger the machine check recovery process.

Signed-off-by: Qiuxu Zhuo <qiuxu.zhuo@intel.com>
Signed-off-by: Tony Luck <tony.luck@intel.com>